quell warning
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@14701 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/3D/FontManager.cxx
@@ -365,7 +365,7 @@ void FontManager::drawString(float x, float y, float z, int faceID, float size,
   }
 
   char buffer[1024];
-  int textlen = strlen(text);
+  int textlen = (int)strlen(text);
 
   assert(textlen < 1024 && "drawString text is way bigger than ever expected");
   memcpy(buffer, text, textlen);
